--- src/main/java/org/tlesis/squakefabric/Reference.java
@@ -1,9 +1,0 @@
-package org.tlesis.squakefabric;
-
-import fi.dy.masa.malilib.util.StringUtils;
-
-public class Reference {
-    public static final String MOD_ID = "squakefabric";
-    public static final String MOD_NAME = "Squake Fabric";
-    public static final String MOD_VERSION = StringUtils.getModVersionString(MOD_ID);
-}
