--- src/main/java/org/tlesis/squakefabric/SquakeFabric.java
@@ -1,16 +1,0 @@
-package org.tlesis.squakefabric;
-
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-import fi.dy.masa.malilib.event.InitializationHandler;
-import net.fabricmc.api.ModInitializer;
-
-public class SquakeFabric implements ModInitializer {
-
-    public static final Logger logger = LogManager.getLogger(Reference.MOD_ID);
-
-    @Override
-    public void onInitialize() {
-        InitializationHandler.getInstance().registerInitializationHandler(new InitHandler());
-    }
-}
