--- src/main/java/org/tlesis/squakefabric/client/SquakeFabricClient.java
@@ -1,12 +1,0 @@
-package org.tlesis.squakefabric.client;
-
-import net.fabricmc.api.ClientModInitializer;
-import net.fabricmc.api.EnvType;
-import net.fabricmc.api.Environment;
-
-@Environment(EnvType.CLIENT)
-public class SquakeFabricClient implements ClientModInitializer {
-
-    @Override
-    public void onInitializeClient() {}
-}
