--- src/main/java/org/tlesis/squakefabric/event/RenderHandler.java
@@ -1,5 +1,0 @@
-package org.tlesis.squakefabric.event;
-
-import fi.dy.masa.malilib.interfaces.IRenderer;
-
-public class RenderHandler implements IRenderer {}
