--- src/main/java/org/tlesis/squakefabric/scheduler/ClientTickHandler.java
@@ -1,20 +1,0 @@
-package org.tlesis.squakefabric.scheduler;
-
-import fi.dy.masa.malilib.interfaces.IClientTickHandler;
-import net.minecraft.client.MinecraftClient;
-
-public class ClientTickHandler implements IClientTickHandler {
-    public static boolean isJumping = false;
-    
-    @Override
-    public void onClientTick(MinecraftClient mc) {
-
-        if (mc.world != null && mc.player != null) {
-            TaskScheduler.getInstanceClient().runTasks();
-        }
-        
-        if (mc.player != null) {
-            isJumping = mc.player.input.jumping;
-        }
-    }
-}
