+++ src/main/java/org/tlesis/squakeplusplus/Reference.java
@@ -1,0 +1,9 @@
+package org.tlesis.squakeplusplus;
+
+import fi.dy.masa.malilib.util.StringUtils;
+
+public class Reference {
+    public static final String MOD_ID = "squakeplusplus";
+    public static final String MOD_NAME = "Squake++";
+    public static final String MOD_VERSION = StringUtils.getModVersionString(MOD_ID);
+}
