+++ src/main/java/org/tlesis/squakeplusplus/client/SquakePlusPlusClient.java
@@ -1,0 +1,12 @@
+package org.tlesis.squakeplusplus.client;
+
+import net.fabricmc.api.ClientModInitializer;
+import net.fabricmc.api.EnvType;
+import net.fabricmc.api.Environment;
+
+@Environment(EnvType.CLIENT)
+public class SquakePlusPlusClient implements ClientModInitializer {
+
+    @Override
+    public void onInitializeClient() {}
+}
