+++ src/main/java/org/tlesis/squakeplusplus/config/FeatureToggle.java
@@ -1,0 +1,197 @@
+package org.tlesis.squakeplusplus.config;
+
+import com.google.common.collect.ImmutableList;
+import com.google.gson.JsonElement;
+import com.google.gson.JsonPrimitive;
+
+import org.tlesis.squakeplusplus.SquakePlusPlus;
+
+import fi.dy.masa.malilib.config.ConfigType;
+import fi.dy.masa.malilib.config.IConfigBoolean;
+import fi.dy.masa.malilib.config.IConfigNotifiable;
+import fi.dy.masa.malilib.config.IHotkeyTogglable;
+import fi.dy.masa.malilib.gui.GuiBase;
+import fi.dy.masa.malilib.hotkeys.IKeybind;
+import fi.dy.masa.malilib.hotkeys.KeyCallbackToggleBooleanConfigWithMessage;
+import fi.dy.masa.malilib.hotkeys.KeybindMulti;
+import fi.dy.masa.malilib.hotkeys.KeybindSettings;
+import fi.dy.masa.malilib.interfaces.IValueChangeCallback;
+import fi.dy.masa.malilib.util.StringUtils;
+
+public enum FeatureToggle implements IHotkeyTogglable, IConfigNotifiable<IConfigBoolean> {
+
+    ENABLED               ("Squake",                    true,  "", "Enables/disables all changes to movement"),
+    BHOP                  ("Bhopping",                  true,  "", "Enables bunnyhopping and airstrafing"),
+    UNCAPPED_BHOP         ("Uncapped Bhopping",         true,  "", "If enabled, the soft and hard speed caps will not be applied at all"),
+    SHARK                 ("Sharking",                  false, "", "Enables sharking"),
+    TRIMP                 ("Trimping",                  false, "", "Enables trimping"),
+    HL2_OLD_ENGINE_BHOP   ("HL2 Old Engine Bhop",       false, "", "Makes Bhopping behave like Half-Life 2 Old Engine bhopping"),
+    JUMP_SPAM             ("Jump Spam",                 true,  "", "Spams jump rather than having the cooldown"),
+    SPEEDOMETER           ("Speedometer",               false, "", "Its just a Speedometer");
+    
+    public static final ImmutableList<FeatureToggle> VALUES = ImmutableList.copyOf(values());
+
+    private final String name;
+    private final String comment;
+    private final String prettyName;
+    private final IKeybind keybind;
+    private final boolean defaultValueBoolean;
+    private final boolean singlePlayer;
+    private boolean valueBoolean;
+    private IValueChangeCallback<IConfigBoolean> callback;
+
+    FeatureToggle(String name, boolean defaultValue, String defaultHotkey, String comment) {
+        this(name, defaultValue, false, defaultHotkey, KeybindSettings.DEFAULT, comment);
+    }
+
+    FeatureToggle(String name, boolean defaultValue, boolean singlePlayer, String defaultHotkey, String comment) {
+        this(name, defaultValue, singlePlayer, defaultHotkey, KeybindSettings.DEFAULT, comment);
+    }
+
+    FeatureToggle(String name, boolean defaultValue, String defaultHotkey, KeybindSettings settings, String comment) {
+        this(name, defaultValue, false, defaultHotkey, settings, comment);
+    }
+
+    FeatureToggle(String name, boolean defaultValue, boolean singlePlayer, String defaultHotkey, KeybindSettings settings, String comment) {
+        this(name, defaultValue, singlePlayer, defaultHotkey, settings, comment, StringUtils.splitCamelCase(name));
+    }
+
+    FeatureToggle(String name, boolean defaultValue, String defaultHotkey, String comment, String prettyName) {
+        this(name, defaultValue, false, defaultHotkey, comment, prettyName);
+    }
+
+    FeatureToggle(String name, boolean defaultValue, boolean singlePlayer, String defaultHotkey, String comment, String prettyName) {
+        this(name, defaultValue, singlePlayer, defaultHotkey, KeybindSettings.DEFAULT, comment, prettyName);
+    }
+
+    FeatureToggle(String name, boolean defaultValue, boolean singlePlayer, String defaultHotkey, KeybindSettings settings, String comment, String prettyName) {
+        this.name = name;
+        this.valueBoolean = defaultValue;
+        this.defaultValueBoolean = defaultValue;
+        this.singlePlayer = singlePlayer;
+        this.comment = comment;
+        this.prettyName = prettyName;
+        this.keybind = KeybindMulti.fromStorageString(defaultHotkey, settings);
+        this.keybind.setCallback(new KeyCallbackToggleBooleanConfigWithMessage(this));
+    }
+
+    @Override
+    public ConfigType getType() {
+        return ConfigType.HOTKEY;
+    }
+
+    @Override
+    public String getName() {
+        return this.name;
+    }
+
+    @Override
+    public String getConfigGuiDisplayName() {
+        String name = StringUtils.getTranslatedOrFallback("config.name." + this.getName().toLowerCase(), this.getName());
+
+        if (this.singlePlayer) {
+            return GuiBase.TXT_GOLD + name + GuiBase.TXT_RST;
+        }
+
+        return name;
+    }
+    
+    @Override
+    public String getPrettyName() {
+        return this.prettyName;
+    }
+
+    @Override
+    public String getStringValue() {
+        return String.valueOf(this.valueBoolean);
+    }
+
+    @Override
+    public String getDefaultStringValue() {
+        return String.valueOf(this.defaultValueBoolean);
+    }
+
+    @Override
+    public void setValueFromString(String value) {}
+
+    @Override
+    public void onValueChanged() {
+        if (this.callback != null) {
+            this.callback.onValueChanged(this);
+        }
+    }
+
+    @Override
+    public void setValueChangeCallback(IValueChangeCallback<IConfigBoolean> callback) {
+        this.callback = callback;
+    }
+
+    @Override
+    public String getComment() {
+        String comment = StringUtils.getTranslatedOrFallback("config.comment." + this.getName().toLowerCase(), this.comment);
+
+        if (comment != null && this.singlePlayer) {
+            return comment + "\n" + StringUtils.translate("tweakeroo.label.config_comment.single_player_only");
+        }
+
+        return comment;
+    }
+
+    @Override
+    public IKeybind getKeybind() {
+        return this.keybind;
+    }
+
+    @Override
+    public boolean getBooleanValue() {
+        return this.valueBoolean;
+    }
+
+    @Override
+    public boolean getDefaultBooleanValue() {
+        return this.defaultValueBoolean;
+    }
+
+    @Override
+    public void setBooleanValue(boolean value) {
+        boolean oldValue = this.valueBoolean;
+        this.valueBoolean = value;
+
+        if (oldValue != this.valueBoolean) {
+            this.onValueChanged();
+        }
+    }
+
+    @Override
+    public boolean isModified() {
+        return this.valueBoolean != this.defaultValueBoolean;
+    }
+
+    @Override
+    public boolean isModified(String newValue) {
+        return Boolean.parseBoolean(newValue) != this.defaultValueBoolean;
+    }
+
+    @Override
+    public void resetToDefault() {
+        this.valueBoolean = this.defaultValueBoolean;
+    }
+
+    @Override
+    public JsonElement getAsJsonElement() {
+        return new JsonPrimitive(this.valueBoolean);
+    }
+
+    @Override
+    public void setValueFromJsonElement(JsonElement element) {
+        try {
+            if (element.isJsonPrimitive()) {
+                this.valueBoolean = element.getAsBoolean();
+            } else {
+                SquakePlusPlus.logger.warn("Failed to set config value for '{}' from the JSON element '{}'", this.getName(), element);
+            }
+        } catch (Exception e) {
+            SquakePlusPlus.logger.warn("Failed to set config value for '{}' from the JSON element '{}'", this.getName(), element, e);
+        }
+    }
+}
