+++ src/main/java/org/tlesis/squakeplusplus/event/RenderHandler.java
@@ -1,0 +1,5 @@
+package org.tlesis.squakeplusplus.event;
+
+import fi.dy.masa.malilib.interfaces.IRenderer;
+
+public class RenderHandler implements IRenderer {}
