+++ src/main/resources/mixins.squake-fabric.json
@@ -1,7 +1,7 @@
 {
   "required": true,
   "minVersion": "0.8",
-  "package": "org.tlesis.squakefabric.mixin",
+  "package": "org.tlesis.squakeplusplus.mixin",
   "compatibilityLevel": "JAVA_17",
   "mixins": [
   ],
